+++ src/main/java/com/example/news_feed/admin/controller/AdminCommentController.java
@@ -1,13 +1,12 @@
 package com.example.news_feed.admin.controller;
 
-import com.example.news_feed.admin.service.AdminCommentService;
+import com.example.news_feed.admin.serviceImpl.AdminCommentServiceImpl;
 import com.example.news_feed.comment.dto.request.UpdateCommentDto;
 import com.example.news_feed.comment.dto.response.CommentDetailDto;
 import com.example.news_feed.comment.dto.response.CommentResponseDto;
-import com.example.news_feed.comment.service.CommentService;
+import com.example.news_feed.comment.serviceImpl.CommentServiceImpl;
 import com.example.news_feed.auth.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -20,8 +19,8 @@ import java.util.List;
 @RequestMapping("/admin/*")
 public class AdminCommentController {
 
-    private final AdminCommentService AdmincommentService;
-    private final CommentService commentService;
+    private final AdminCommentServiceImpl admincommentServiceImpl;
+    private final CommentServiceImpl commentServiceImpl;
 
     // 댓글 수정
     @PatchMapping("/comments/{commentId}")
@@ -29,7 +28,7 @@ public class AdminCommentController {
                                                              @RequestBody UpdateCommentDto updateCommentDto,
                                                              @AuthenticationPrincipal final UserDetailsImpl userDetails) {
         updateCommentDto.setUserId(userDetails.getId());
-        commentService.update(commentId, updateCommentDto);
+        commentServiceImpl.update(commentId, updateCommentDto);
 
         CommentResponseDto response = CommentResponseDto.res(HttpStatus.OK.value(), "댓글 수정 완료");
         return ResponseEntity.status(HttpStatus.OK).body(response);
@@ -40,7 +39,7 @@ public class AdminCommentController {
     public ResponseEntity<CommentResponseDto> deleteComment(@PathVariable Long commentId,
                                                             @AuthenticationPrincipal final UserDetailsImpl userDetails){
         Long userId = userDetails.getId();
-        commentService.delete(userId, commentId);
+        commentServiceImpl.delete(userId, commentId);
         CommentResponseDto response =  CommentResponseDto.res(HttpStatus.OK.value(), "댓글 삭제 완료");
         return ResponseEntity.status(HttpStatus.OK).body(response);
     }
@@ -48,14 +47,14 @@ public class AdminCommentController {
     // 특정 게시글의 댓글 목록 조회
     @GetMapping("/post/{postId}/comment")
     public ResponseEntity<List<CommentDetailDto>> showAll(@PathVariable Long postId){
-        List<CommentDetailDto> comments = commentService.showAll(postId);
+        List<CommentDetailDto> comments = commentServiceImpl.showAll(postId);
         return ResponseEntity.status(HttpStatus.OK).body(comments);
     }
 
     // 전체 댓글 목록 조회
     @GetMapping("/post/comment")
     public ResponseEntity<List<CommentDetailDto>> showAllComment(){
-        List<CommentDetailDto> comments = AdmincommentService.showAll();
+        List<CommentDetailDto> comments = admincommentServiceImpl.showAll();
         return ResponseEntity.status(HttpStatus.OK).body(comments);
     }
