+++ src/main/java/com/example/news_feed/admin/controller/AdminPostController.java
@@ -1,14 +1,12 @@
 package com.example.news_feed.admin.controller;
 
-import com.example.news_feed.admin.service.AdminPostService;
-import com.example.news_feed.multimedia.service.MultiMediaService;
+import com.example.news_feed.multimedia.serviceImpl.MultiMediaServiceImpl;
 import com.example.news_feed.post.dto.request.UpdatePostDto;
 import com.example.news_feed.post.dto.response.PostDetailDto;
 import com.example.news_feed.post.dto.response.PostResponseDto;
-import com.example.news_feed.post.service.PostService;
+import com.example.news_feed.post.serviceImpl.PostServiceImpl;
 import com.example.news_feed.auth.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -22,8 +20,8 @@ import java.util.List;
 @RequestMapping("/admin/*")
 public class AdminPostController {
 
-    private final PostService postService;
-    private final MultiMediaService multiMediaService;
+    private final PostServiceImpl postServiceImpl;
+    private final MultiMediaServiceImpl multiMediaServiceImpl;
 
 
     // 게시글 수정
@@ -32,7 +30,7 @@ public class AdminPostController {
                                                   @RequestBody UpdatePostDto updatePostDto,
                                                   @AuthenticationPrincipal final UserDetailsImpl userDetails) {
         updatePostDto.setUserId(userDetails.getId());
-        postService.update(postId, updatePostDto);
+        postServiceImpl.update(postId, updatePostDto);
         PostResponseDto response = PostResponseDto.res(HttpStatus.OK.value(), "게시물 수정 완료");
         return ResponseEntity.status(HttpStatus.OK).body(response);
     }
@@ -42,7 +40,7 @@ public class AdminPostController {
     public ResponseEntity<PostResponseDto> deletePost(@PathVariable Long postId,
                                                   @AuthenticationPrincipal final UserDetailsImpl userDetails) {
         Long userId = userDetails.getId();
-        postService.delete(userId, postId);
+        postServiceImpl.delete(userId, postId);
         PostResponseDto response = PostResponseDto.res(HttpStatus.OK.value(), "게시물 삭제 완료");
         return ResponseEntity.status(HttpStatus.OK).body(response);
     }
@@ -50,7 +48,7 @@ public class AdminPostController {
     // 게시글 목록 조회
     @GetMapping("/post")
     public ResponseEntity<List<PostDetailDto>> showAll(){
-        List<PostDetailDto> posts = postService.showAll();
+        List<PostDetailDto> posts = postServiceImpl.showAll();
         return ResponseEntity.status(HttpStatus.OK).body(posts);
     }
 
@@ -59,7 +57,7 @@ public class AdminPostController {
     public ResponseEntity<PostResponseDto> updateFile(@PathVariable Long postId,
                                                       @RequestPart(value ="files", required = false) List<MultipartFile> files) {
 
-        multiMediaService.updateFile(postId, files);
+        multiMediaServiceImpl.updateFile(postId, files);
 
         PostResponseDto response = PostResponseDto.res(HttpStatus.OK.value(), "게시물 수정 완료");
         return ResponseEntity.status(HttpStatus.OK).body(response);
@@ -70,7 +68,7 @@ public class AdminPostController {
     public ResponseEntity<PostResponseDto> deleteFile(@PathVariable Long multiMediaId,
                                                       @AuthenticationPrincipal final UserDetailsImpl userDetails) {
         Long userId = userDetails.getId();
-        multiMediaService.deleteFiles(userId, multiMediaId);
+        multiMediaServiceImpl.deleteFiles(userId, multiMediaId);
         PostResponseDto response = PostResponseDto.res(HttpStatus.OK.value(), "게시물 삭제 완료");
         return ResponseEntity.status(HttpStatus.OK).body(response);
     }
