+++ src/main/java/com/example/news_feed/admin/service/AdminUserService.java
@@ -1,65 +1,30 @@
 package com.example.news_feed.admin.service;
 
 import com.example.news_feed.admin.dto.response.UserDetailDto;
-import com.example.news_feed.admin.repository.AdminUserRepository;
-import com.example.news_feed.common.exception.HttpException;
-import com.example.news_feed.user.domain.User;
 import com.example.news_feed.user.domain.UserRoleEnum;
-import com.example.news_feed.user.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
-@Service
-public class AdminUserService {
+public interface AdminUserService {
+    /*
+     * 유저 목록
+     * @return 유저 전체 목록
+     */
+    List<UserDetailDto> showAll();
 
-    @Autowired
-    private AdminUserRepository adminUserRepository;
+    /*
+     * 유저 권한 변경
+     * @param newRole 변경하려는 권한 정보
+     * @param userId 대상 유저 정보
+     * @return 유저 권한 변경 결과
+     */
+    void changeUserRole(Long userId, UserRoleEnum newRole);
 
-    @Autowired
-    private UserRepository userRepository;
-
-    // 유저 전체 목록(탈퇴회원 포함)
-    public List<UserDetailDto> showAll() {
-
-        return adminUserRepository.findAll()
-                .stream()
-                .map(UserDetailDto::createUserDetailDto)
-                .collect(Collectors.toList());
-    }
-
-    // 유저 권한 변경
-    public void changeUserRole(Long userId, UserRoleEnum newRole) {
-        // 기존 유저정보 조회 및 예외 처리
-        User target = checkUser(userId);
-
-        // 같은 권한으로 변경하려 할 경우
-        if (target.getRole().equals(newRole)) {
-            throw new HttpException("이미 해당 권한을 가지고 있습니다.", HttpStatus.BAD_REQUEST);
-        }
-
-        target.setRole(newRole);
-        userRepository.save(target);
-    }
-
-    // 회원 상태 변경(탈퇴, 복구)
-    public void changeStatus(Long userId, String status) {
-        // 기존 유저정보 조회 및 예외 처리
-        User target = checkUser(userId);
-
-        if (target.getStatus().equals(status)) {
-            throw new HttpException("변경할 상태가 없습니다.", HttpStatus.BAD_REQUEST);
-        }
-        target.setStatus(status);
-        userRepository.save(target);
-    }
-
-    // 유저 정보 확인
-    private User checkUser(Long userId) {
-        return userRepository.findById(userId)
-                .orElseThrow(() -> new HttpException("유저 정보가 없습니다.", HttpStatus.BAD_REQUEST));
-    }
+    /*
+     * 유저 상태 변경(탈퇴)
+     * @param status 변경하려는 상태 정보
+     * @param userId 대상 유저 정보
+     * @return 유저 상태 변경 결과
+     */
+    void changeStatus(Long userId, String status);
 }
