+++ src/main/java/com/example/news_feed/comment/controller/CommentApiController.java
@@ -4,11 +4,10 @@ import com.example.news_feed.comment.dto.request.CreateCommentDto;
 import com.example.news_feed.comment.dto.request.UpdateCommentDto;
 import com.example.news_feed.comment.dto.response.CommentDetailDto;
 import com.example.news_feed.comment.dto.response.CommentResponseDto;
-import com.example.news_feed.comment.service.CommentService;
+import com.example.news_feed.comment.serviceImpl.CommentServiceImpl;
 import com.example.news_feed.auth.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -21,7 +20,7 @@ import java.util.List;
 @Slf4j
 public class CommentApiController {
 
-    private final CommentService commentService;
+    private final CommentServiceImpl commentServiceImpl;
 
     // 댓글 등록
     @PostMapping("/post/{postId}/comment")
@@ -30,7 +29,7 @@ public class CommentApiController {
                                                       @AuthenticationPrincipal final UserDetailsImpl userDetails) {
         createCommentDto.setUserId(userDetails.getId());
 
-        commentService.create(postId, createCommentDto);
+        commentServiceImpl.create(postId, createCommentDto);
         CommentResponseDto response = CommentResponseDto.res(HttpStatus.CREATED.value(), "댓글 등록 완료");
         return ResponseEntity.status(HttpStatus.CREATED).body(response);
     }
@@ -42,7 +41,7 @@ public class CommentApiController {
                                                       @AuthenticationPrincipal final UserDetailsImpl userDetails) {
         updateCommentDto.setUserId(userDetails.getId());
         log.info("Controller userId {}", userDetails.getId());
-        commentService.update(commentId, updateCommentDto);
+        commentServiceImpl.update(commentId, updateCommentDto);
 
         CommentResponseDto response = CommentResponseDto.res(HttpStatus.OK.value(), "댓글 수정 완료");
         return ResponseEntity.status(HttpStatus.OK).body(response);
@@ -53,7 +52,7 @@ public class CommentApiController {
     public ResponseEntity<CommentResponseDto> delete(@PathVariable Long commentId,
                                                      @AuthenticationPrincipal final UserDetailsImpl userDetails){
         Long userId = userDetails.getId();
-        commentService.delete(userId, commentId);
+        commentServiceImpl.delete(userId, commentId);
         CommentResponseDto response =  CommentResponseDto.res(HttpStatus.OK.value(), "댓글 삭제 완료");
         return ResponseEntity.status(HttpStatus.OK).body(response);
     }
@@ -61,7 +60,7 @@ public class CommentApiController {
     // 댓글 목록
     @GetMapping("/post/{postId}/comment")
     public ResponseEntity<List<CommentDetailDto>> showAll(@PathVariable Long postId){
-        List<CommentDetailDto> comments = commentService.showAll(postId);
+        List<CommentDetailDto> comments = commentServiceImpl.showAll(postId);
         return ResponseEntity.status(HttpStatus.OK).body(comments);
     }
 
@@ -69,6 +68,6 @@ public class CommentApiController {
     // 댓글 상세 조회
     @GetMapping("/comments/{commentId}")
     public ResponseEntity<CommentDetailDto> show(@PathVariable Long commentId){
-        return ResponseEntity.status(HttpStatus.OK).body(commentService.show(commentId));
+        return ResponseEntity.status(HttpStatus.OK).body(commentServiceImpl.show(commentId));
     }
 }
