+++ src/main/java/com/example/news_feed/comment/controller/CommentLikeApiController.java
@@ -1,11 +1,10 @@
 package com.example.news_feed.comment.controller;
 
 import com.example.news_feed.comment.dto.response.CommentResponseDto;
-import com.example.news_feed.comment.service.CommentLikeService;
+import com.example.news_feed.comment.serviceImpl.CommentLikeServiceImpl;
 import com.example.news_feed.auth.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -19,7 +18,7 @@ import org.springframework.web.bind.annotation.RestController;
 @Slf4j
 public class CommentLikeApiController {
 
-    private final CommentLikeService commentLikeService;
+    private final CommentLikeServiceImpl commentLikeServiceImpl;
 
     // 좋아요 등록
     @PostMapping("/comment/{commentId}/like")
@@ -27,18 +26,18 @@ public class CommentLikeApiController {
                                                      @AuthenticationPrincipal final UserDetailsImpl userDetails){
 
         Long userId = userDetails.getId();
-        commentLikeService.create(commentId, userId);
+        commentLikeServiceImpl.create(commentId, userId);
         CommentResponseDto response = CommentResponseDto.res(HttpStatus.CREATED.value(), "댓글 좋아요 +1");
         return ResponseEntity.status(HttpStatus.CREATED).body(response);
     }
 
     // 좋아요 취소
-    @DeleteMapping("/comment/{commentId}/like")
-    public ResponseEntity<CommentResponseDto> delete(@PathVariable Long commentId,
+    @DeleteMapping("/comment/{commentLikeId}/like")
+    public ResponseEntity<CommentResponseDto> delete(@PathVariable Long commentLikeId,
                                                      @AuthenticationPrincipal final UserDetailsImpl userDetails){
 
         Long userId = userDetails.getId();
-        commentLikeService.delete(commentId, userId);
+        commentLikeServiceImpl.delete(commentLikeId, userId);
         CommentResponseDto response = CommentResponseDto.res(HttpStatus.OK.value(), "댓글 좋아요 취소");
         return ResponseEntity.status(HttpStatus.OK).body(response);
     }
