+++ src/main/java/com/example/news_feed/post/controller/PostApiController.java
@@ -1,13 +1,12 @@
 package com.example.news_feed.post.controller;
 
-import com.example.news_feed.common.aop.annotation.RunningTime;
 import com.example.news_feed.multimedia.dto.MultiMediaDto;
-import com.example.news_feed.multimedia.service.MultiMediaService;
+import com.example.news_feed.multimedia.serviceImpl.MultiMediaServiceImpl;
 import com.example.news_feed.post.dto.request.CreatePostDto;
 import com.example.news_feed.post.dto.request.UpdatePostDto;
 import com.example.news_feed.post.dto.response.PostDetailDto;
 import com.example.news_feed.post.dto.response.PostResponseDto;
-import com.example.news_feed.post.service.PostService;
+import com.example.news_feed.post.serviceImpl.PostServiceImpl;
 import com.example.news_feed.auth.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -24,8 +23,8 @@ import java.util.List;
 @Slf4j
 public class PostApiController {
 
-    private final PostService postService;
-    private final MultiMediaService multiMediaService;
+    private final PostServiceImpl postServiceImpl;
+    private final MultiMediaServiceImpl multiMediaServiceImpl;
 
     // 게시글 작성
     @PostMapping("/post")
@@ -33,7 +32,7 @@ public class PostApiController {
                                                   @AuthenticationPrincipal final UserDetailsImpl userDetails) {
 
         createPostDto.setUserId(userDetails.getId());
-        postService.createPost(createPostDto);
+        postServiceImpl.createPost(createPostDto);
         PostResponseDto response = PostResponseDto.res(HttpStatus.CREATED.value(), "게시물 등록 완료");
         return ResponseEntity.status(HttpStatus.CREATED).body(response);
     }
@@ -42,7 +41,7 @@ public class PostApiController {
     @PostMapping("/post/{postId}/file")
     public ResponseEntity<PostResponseDto> createFile(@RequestPart(value="files", required = false) List<MultipartFile> files,
                                                       @PathVariable Long postId) {
-        multiMediaService.createFile(files, postId);
+        multiMediaServiceImpl.createFile(files, postId);
         PostResponseDto response = PostResponseDto.res(HttpStatus.CREATED.value(), "게시물 등록 완료");
         return ResponseEntity.status(HttpStatus.CREATED).body(response);
     }
@@ -54,7 +53,7 @@ public class PostApiController {
                                                   @AuthenticationPrincipal final UserDetailsImpl userDetails) {
 
         updatePostDto.setUserId(userDetails.getId());
-        postService.update(postId, updatePostDto);
+        postServiceImpl.update(postId, updatePostDto);
         PostResponseDto response = PostResponseDto.res(HttpStatus.OK.value(), "게시물 수정 완료");
         return ResponseEntity.status(HttpStatus.OK).body(response);
     }
@@ -64,7 +63,7 @@ public class PostApiController {
     public ResponseEntity<PostResponseDto> updateFile(@PathVariable Long postId,
                                                   @RequestPart(value ="files", required = false) List<MultipartFile> files) {
 
-        multiMediaService.updateFile(postId, files);
+        multiMediaServiceImpl.updateFile(postId, files);
 
         PostResponseDto response = PostResponseDto.res(HttpStatus.OK.value(), "게시물 수정 완료");
         return ResponseEntity.status(HttpStatus.OK).body(response);
@@ -75,7 +74,7 @@ public class PostApiController {
     public ResponseEntity<PostResponseDto> delete(@PathVariable Long postId,
                                        @AuthenticationPrincipal final UserDetailsImpl userDetails) {
         Long userId = userDetails.getId();
-        postService.delete(userId, postId);
+        postServiceImpl.delete(userId, postId);
         PostResponseDto response = PostResponseDto.res(HttpStatus.OK.value(), "게시물 삭제 완료");
         return ResponseEntity.status(HttpStatus.OK).body(response);
     }
@@ -85,7 +84,7 @@ public class PostApiController {
     public ResponseEntity<PostResponseDto> deleteFile(@PathVariable Long multiMediaId,
                                                       @AuthenticationPrincipal final UserDetailsImpl userDetails) {
         Long userId = userDetails.getId();
-        multiMediaService.deleteFiles(userId, multiMediaId);
+        multiMediaServiceImpl.deleteFiles(userId, multiMediaId);
         PostResponseDto response = PostResponseDto.res(HttpStatus.OK.value(), "게시물 삭제 완료");
         return ResponseEntity.status(HttpStatus.OK).body(response);
     }
@@ -95,7 +94,7 @@ public class PostApiController {
 //    @RunningTime
     @GetMapping("/post")
     public ResponseEntity<List<PostDetailDto>> showAll(){
-        List<PostDetailDto> posts = postService.showAll();
+        List<PostDetailDto> posts = postServiceImpl.showAll();
         return ResponseEntity.status(HttpStatus.OK).body(posts);
     }
 
@@ -103,13 +102,13 @@ public class PostApiController {
     // 게시글 상세
     @GetMapping("/post/{postId}")
     public ResponseEntity<PostDetailDto> show(@PathVariable Long postId){
-        return ResponseEntity.status(HttpStatus.OK).body(postService.show(postId));
+        return ResponseEntity.status(HttpStatus.OK).body(postServiceImpl.show(postId));
     }
 
     // 멀티미디어 조회
     @GetMapping("/post/{postId}/file")
     public ResponseEntity<List<MultiMediaDto>> showFile(@PathVariable Long postId){
-        return ResponseEntity.status(HttpStatus.OK).body(multiMediaService.showFiles(postId));
+        return ResponseEntity.status(HttpStatus.OK).body(multiMediaServiceImpl.showFiles(postId));
     }
 
 }
