+++ src/main/java/com/example/news_feed/post/controller/PostLikeApiController.java
@@ -1,10 +1,9 @@
 package com.example.news_feed.post.controller;
 
 import com.example.news_feed.post.dto.response.PostResponseDto;
-import com.example.news_feed.post.service.PostLikeService;
+import com.example.news_feed.post.serviceImpl.PostLikeServiceImpl;
 import com.example.news_feed.auth.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -14,7 +13,7 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class PostLikeApiController {
 
-    private final PostLikeService postLikeService;
+    private final PostLikeServiceImpl postLikeServiceImpl;
 
     // 좋아요 등록
     @PostMapping("/post/{postId}/like")
@@ -22,18 +21,18 @@ public class PostLikeApiController {
                                                   @AuthenticationPrincipal final UserDetailsImpl userDetails){
 
         Long userId = userDetails.getId();
-        postLikeService.create(postId, userId);
+        postLikeServiceImpl.create(postId, userId);
         PostResponseDto response = PostResponseDto.res(HttpStatus.CREATED.value(), "게시글 좋아요 +1");
         return ResponseEntity.status(HttpStatus.CREATED).body(response);
     }
 
     // 좋아요 취소
-    @DeleteMapping("/post/{postId}/like")
-    public ResponseEntity<PostResponseDto> delete(@PathVariable Long postId,
+    @DeleteMapping("/post/{postLikeId}/like")
+    public ResponseEntity<PostResponseDto> delete(@PathVariable Long postLikeId,
                                                   @AuthenticationPrincipal final UserDetailsImpl userDetails) {
 
         Long userId = userDetails.getId();
-        postLikeService.delete(userId, postId);
+        postLikeServiceImpl.delete(userId, postLikeId);
         PostResponseDto response = PostResponseDto.res(HttpStatus.OK.value(), "게시글 좋아요 취소");
         return ResponseEntity.status(HttpStatus.OK).body(response);
