+++ src/main/java/com/example/news_feed/user/controller/NavigateController.java
@@ -2,7 +2,7 @@ package com.example.news_feed.user.controller;
 
 import com.example.news_feed.admin.dto.response.UserDetailDto;
 import com.example.news_feed.auth.security.UserDetailsImpl;
-import com.example.news_feed.user.service.MypageService;
+import com.example.news_feed.user.serviceImpl.MyPageServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 @RequiredArgsConstructor
 public class NavigateController {
 
-    private final MypageService mypageService;
+    private final MyPageServiceImpl mypageServiceImpl;
 
     // 회원가입 페이지 이동
     @GetMapping("/signup-page")
@@ -25,7 +25,7 @@ public class NavigateController {
     @GetMapping("/home")
     public String home(Model model, @AuthenticationPrincipal UserDetailsImpl userDetails) {
         if (userDetails != null){
-            UserDetailDto user = mypageService.showUser(userDetails.getId());
+            UserDetailDto user = mypageServiceImpl.showUser(userDetails.getId());
             model.addAttribute("user", user);
         }
         return "main/home";
